+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/log2prov/interfaces/ProxyMakerInterface.java
@@ -5,11 +5,12 @@ import org.openprovenance.prov.model.QualifiedName;
 import java.util.Collection;
 import java.util.Map;
 import java.util.Set;
+import java.util.function.Function;
 
 // This interface is useful to invoke method on generated classes, by means of the ProxyManagement class, without having to share any package/classes.
 public interface ProxyMakerInterface {
     Object make(Object[] record, Object _processor);   // public <T> T make(Object[] record, Template_blockBuilderInterface<T> _processor) {
-    Object getTypeManager(Map<QualifiedName, Set<String>> ktm, Map<QualifiedName, Set<String>> utm);
+    Object getTypeManager(Map<QualifiedName, Set<String>> ktm, Map<QualifiedName, Set<String>> utm, Map<String, Map<String, Function<Object, String>>> propertyConverters);
     Object getTypedRecord();
     void propagateTypes(Object[] record,
                         Map<String, Integer> mapLevelN,