+++ modules-template/prov-template-compiler/src/main/resources/org/openprovenance/prov/template/compiler/script.sh
@@ -61,6 +61,16 @@ while [[ $# -gt 0 ]]; do
           shift # past argument
           shift # past value
           ;;
+    -addLevel0|--addLevel0)
+          _ADD_LEVEL0="$2"
+          shift # past argument
+          shift # past value
+          ;;
+    -propertyConverters|--propertyConverters)
+          _PROPERTY_CONVERTERS="$2"
+          shift # past argument
+          shift # past value
+          ;;
     --default)
       export DEFAULT=YES
       shift # past argument
@@ -84,7 +94,7 @@ fi
 
 
 if [ "$ARG1" = "kernel" ]; then
-   (export LEVEL_NUMBER="$_LEVEL_NUMBER" TRANSLATION="$_TRANSLATION" KNOWN_TYPES="$_KNOWN_TYPES" KNOWN_RELATIONS="$_KNOWN_RELATIONS" RELATION_OFFSET="$_RELATION_OFFSET"; export LEVEL_OFFSET="$_LEVEL_OFFSET"; export SET_OFFSET="$_SET_OFFSET"; $PROVCONVERT -infile $_INFILE -log2prov ${INIT} -outfile $_OUTFILE -log2kernel)
+   (export PROPERTY_CONVERTERS="$_PROPERTY_CONVERTERS" ADD_LEVEL0="$_ADD_LEVEL0" LEVEL_NUMBER="$_LEVEL_NUMBER" TRANSLATION="$_TRANSLATION" KNOWN_TYPES="$_KNOWN_TYPES" KNOWN_RELATIONS="$_KNOWN_RELATIONS" RELATION_OFFSET="$_RELATION_OFFSET"; export LEVEL_OFFSET="$_LEVEL_OFFSET"; export SET_OFFSET="$_SET_OFFSET"; $PROVCONVERT -infile $_INFILE -log2prov ${INIT} -outfile $_OUTFILE -log2kernel)
 else
    $PROVCONVERT -infile "$_INFILE" -log2prov ${INIT} -outfile "$_OUTFILE"
 fi